Add firstboot service for config
Add a new service which runs on boot and configures the system based
upon values in /boot/pikvm.txt . If that file exists, the service mounts
filesystems as read-write, applies configuration, and then deletes the
file.

pikvm.txt is a sourced shell script, but it is intended to be a file in
.env format. It accepts a subset of the variables defined in the
config.mk .

Currently, the accepted variables are WIFI_ESSID, WIFI_PASSWD, and
WIFI_IFACE. Like the build scripts, network configuration is only
written if WIFI_ESSID is supplied. WIFI_IFACE defaults to "wlan0" like
in the build scripts if not specified. Note that if there is previous
configuration for that SSID, it is completely ovewritten and not merged.